+++ pytorch_lightning/callbacks/__init__.py
@@ -22,7 +22,7 @@ from pytorch_lightning.callbacks.model_checkpoint import ModelCheckpoint
 from pytorch_lightning.callbacks.progress import ProgressBar, ProgressBarBase
 from pytorch_lightning.callbacks.pruning import ModelPruning
 from pytorch_lightning.callbacks.quantization import QuantizationAwareTraining
-from pytorch_lightning.callbacks.swa import StochasticWeightAveraging
+from pytorch_lightning.callbacks.stochastic_weight_avg import StochasticWeightAveraging
 
 __all__ = [
     'BackboneFinetuning',
