+++ pytorch_lightning/callbacks/stochastic_weight_avg.py
