+++ tests/callbacks/test_stochastic_weight_avg.py
